20/03/2022

+++ Ascensor.WebAPI/DTO/Entities/AscensorEntity.cs
@@ -5,7 +5,7 @@
         public int Asce_Id { get; set; }
         public int Asce_Piso { get; set; }
         public bool Asce_MiUbicacion { get; set; }
-        public decimal Asce_Tiempo { get; set; }
+        public int Asce_Tiempo { get; set; }
         public bool Asce_Estado { get; set; }
     }
 }
